requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
pytest
pytest-flask
xhtml2pdf
pytesseract
Pillow